refactor: 탭 Tab 컴포넌트에서 context를 관리하도록 수정
